+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/iOS/ContextMenuPlatformEffect.Pressed.cs
@@ -15,7 +15,7 @@ public partial class ContextMenuPlatformEffect
     private void SetupPressedMode(ContextMenu contextMenu)
     {
         contextMenu.ItemsSourceUpdated += RebuildMenu;
-        contextMenu.ItemsPropertiesUpdated += RebuildMenu;
+        contextMenu.ItemPropertiesUpdated += RebuildMenu;
         
         if (Control is not UIButton uiButton)
         {
@@ -98,7 +98,7 @@ public partial class ContextMenuPlatformEffect
         m_uiButtonToRemove?.RemoveFromSuperview();
         
         m_contextMenu.ItemsSourceUpdated -= RebuildMenu;
-        m_contextMenu.ItemsPropertiesUpdated -= RebuildMenu;
+        m_contextMenu.ItemPropertiesUpdated -= RebuildMenu;
         
         if(Element is not null)
             Element.PropertyChanged -= ElementOnPropertyChanged;