+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/ContextMenuItem.Properties.cs
@@ -161,6 +161,11 @@ public partial class ContextMenuItem
         get => (bool)GetValue(ShouldSendGlobalClickProperty);
         set => SetValue(ShouldSendGlobalClickProperty, value);
     }
+    
+    /// <summary>
+    /// The ContextMenu that the item is in
+    /// </summary>
+    public ContextMenu? ContextMenu { get; set; }
 }
 
 /// <summary>